test(vercel-ai-sdk): add otel parity coverage
