+++ packages/vercel-ai-sdk/package.json
@@ -43,6 +43,7 @@
     "ai": ">=7.0.0-beta.111 <8"
   },
   "devDependencies": {
+    "@ai-sdk/otel": "1.0.0-beta.57",
     "ai": "7.0.0-beta.111"
   }
 }
